delete movie

+++ components/admin/DataTable.tsx
@@ -12,7 +12,10 @@ const DataTable: React.FC<DataProps> = ({data, key}) => {
     const router = useRouter()
     const deleteMovie = async (id: string) => {
       try {
+        if(id) {
           await axios.delete(`/api/movies/${id}`)
+          alert('delete sucess')
+        }
       } catch (error) {
           console.log(error);
           return new Response('Error')
@@ -32,10 +35,10 @@ const DataTable: React.FC<DataProps> = ({data, key}) => {
             <td className="whitespace-normal px-4 py-2 text-gray-700">{data.genre}</td>
             <td className="whitespace-normal px-4 py-2 flex gap-4 justify-center">
           <button className="inline-block rounded bg-indigo-600 px-4 py-2 text-xs font-medium text-white hover:bg-red-700">
-          <IoMdAddCircle size={20}/>
+          <IoMdAddCircle size={20} onClick={() => router.push('/admin/addMovies')}/>
           </button>
           <button className="inline-block rounded bg-indigo-600 px-4 py-2 text-xs font-medium text-white hover:bg-red-700">
-            <AiFillEdit size={20} onClick={() => router.push('/admin/addMovies')}/>
+            <AiFillEdit size={20}/>
           </button>
           <button className="inline-block rounded bg-indigo-600 px-4 py-2 text-xs font-medium text-white hover:bg-red-700">
             <AiFillDelete size={20} onClick={() => deleteMovie(data.id)}/>
